requirements.txt
pytest
pytest-pycodestyle
flask
pymongo
pymodm
requests
requests-mock
dnspython
Pillow
numpy
matplotlib
scipy
